Hadi Reviews Changes
