+++ src/components/Shared/PageError.js
@@ -15,7 +15,7 @@ const PageError = ({ onClose, message, height = '' }) => {
       : message?.response?.data
 
   return (
-    <Window Title='Error' width={450} height={height} onClose={onClose} canExpand={false} controlled={true}>
+    <Window Title='Error' width={450} height={height} onClose={onClose} expandable={false} controlled={true}>
       <DialogContent>
         <Box
           sx={{
